+++ CHANGELOG.md
@@ -7,7 +7,10 @@ All notable changes to this project will be documented in this file.
 - Make it easier to manipulate svg files for Imagick and Gmagick by change file name to jpg
 - Fix a bug on save file when we have multiple instances from a model
 - Bugfix on set LARUPLOAD_NULL
-- store human readable file type into database
+- Store human readable file type into database
+- Change type of columns in database
+- Return null for `meta` when value is LARUPLOAD_NULL
+- Streaming - HLS
 
 ## v0.0.2 - 2018/09/10
 - Support Laravel 5.7